course entity

+++ src/main/java/edu/mum/mumsched/courses/controller/CourseController.java
@@ -3,12 +3,16 @@ package edu.mum.mumsched.courses.controller;
 import edu.mum.mumsched.courses.entity.Course;
 import edu.mum.mumsched.courses.service.CourseServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
+import java.util.Collection;
 import java.util.List;
 
-@RestController
+@Controller
 public class CourseController {
 
     @Autowired
@@ -16,28 +20,41 @@ public class CourseController {
 
     @RequestMapping(value ="/courses", method = RequestMethod.GET)
     public String getAllCourses(Model model){
-        List<Course> courses = courseService.getAllCourses();
+        Collection<Course> courses = courseService.getAllCourses();
         model.addAttribute("courses", courses);
         return "courses/view";
     }
 
-    @RequestMapping(value ="/courses/{id}", method = RequestMethod.GET)
-    public Course getCourse(@PathVariable int id){
-        return courseService.getCourse(id);
+    @RequestMapping(value ="/courses/create", method = RequestMethod.GET)
+    public String getCourse(Model model) {
+        model.addAttribute("newcourse", new Course());
+        return "courses/create";
     }
 
     @RequestMapping(value ="/courses", method = RequestMethod.POST)
-    public void createCourse(@RequestBody Course course){
-        courseService.addCourse(course);
+    public String createCourse(Model model, @RequestBody Course course){
+       //Course newCourse = courseService.addCourse(course);
+       //model.addAttribute("newCourse", newCourse);
+       return  "redirect:/courses";
     }
 
     @RequestMapping(value ="/courses/{id}", method = RequestMethod.PUT)
-    public void editCourse(@PathVariable int id, @RequestBody Course course){
+    public String editCourse(@ModelAttribute("updateCourse") @Valid Course course,
+                             BindingResult result, Model model, @PathVariable int id){
+        if(result.hasErrors())
+            return "courses/edit";
+
         courseService.updateCourse(id, course);
+        return "redirect:/courses/";
     }
 
     @RequestMapping(value ="/courses/{id}", method = RequestMethod.DELETE)
-    public void deleteCourse(@PathVariable int id){
+    public String deleteCourse(@PathVariable int id,
+                               BindingResult result, Model model){
+        if(result.hasErrors())
+            return "courses/view";
+
         courseService.deleteCourse(id);
+        return  "redirect: /courses";
     }
 }
